[FE][UI]styling profile

+++ client/src/Components/Album/Album.tsx
@@ -1,19 +1,26 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
+import { FiEdit } from 'react-icons/fi';
+import { MdOutlineExpandMore, MdOutlineExpandLess } from 'react-icons/md';
+import { RiShareForwardLine } from 'react-icons/ri';
 import { TiDelete } from 'react-icons/ti';
+import { useNavigate } from 'react-router-dom';
 import { deleteAlbum } from '../../Services/Server-Client';
-import { AlbumProps } from '../../types/Album.Interfact';
-import { AlbumInterface } from '../../types/Canvas.interface';
-import { User } from '../../types/ServerClient.interface';
+import { AlbumProps } from '../../Types/Album.Interfact';
+import { AlbumInterface } from '../../Types/Canvas.interface';
+import { User } from '../../Types/ServerClient.interface';
 import './Album.css';
 
 const Album = ({ element, editAlbum, setUser }: AlbumProps) => {
+  const navigate = useNavigate();
   const [visibilityOnTitle, setVisibilityOnTitle] = useState<boolean>(false);
+  const [toolToggle, setTooToggle] = useState<boolean>(false);
 
   const handleMouseEnter = () => {
     setVisibilityOnTitle(true);
   };
   const handleMouseLeave = () => {
     setVisibilityOnTitle(false);
+    setTooToggle(false);
   };
 
   const deleteOneAlbum = async function (album: AlbumInterface) {
@@ -38,26 +45,56 @@ const Album = ({ element, editAlbum, setUser }: AlbumProps) => {
         className='albumFrontImage'
         alt='album'
         src={element.frontPage}
-        onClick={() => editAlbum(element)}
+        onClick={() => navigate(`/album/${element.id}/view`)}
       />
+
       {visibilityOnTitle && (
         <>
           <h4 unselectable='on' className='albumTitle'>
             {element.title}
           </h4>
-          {/* <button
-            className='deleteButton'
-            onClick={() => {
-              return deleteOneAlbum(element);
-            }}
-          > */}
-          <TiDelete
-            className=' deleteButton'
-            onClick={() => {
-              return deleteOneAlbum(element);
-            }}
-          />
-          {/* </button> */}
+          {!toolToggle ? (
+            <MdOutlineExpandMore
+              className=' toolButton'
+              onClick={() => setTooToggle(true)}
+            />
+          ) : (
+            <>
+              <MdOutlineExpandLess
+                className=' toolButton'
+                onClick={() => setTooToggle(false)}
+              />
+              <div className='albumTools'>
+                <FiEdit
+                  className='toolIcon'
+                  onClick={() => editAlbum(element)}
+                />
+                <RiShareForwardLine
+                  className='toolIcon'
+                  onClick={async () => {
+                    await navigator.clipboard.writeText(
+                      encodeURI(
+                        `http://localhost:3000/album/${element.id}/view`
+                      )
+                    );
+                    alert('Album link copied 🎉');
+                  }}
+                />
+                <TiDelete
+                  className='toolIcon'
+                  onClick={() => {
+                    if (
+                      window.confirm(
+                        'Are you sure you want to delete this album?'
+                      )
+                    ) {
+                      deleteOneAlbum(element);
+                    }
+                  }}
+                />
+              </div>
+            </>
+          )}
         </>
       )}
     </div>
